Updated settings.rsp to generate code for SDL_guid.h and SDL_joystick.h
SDL_Event has been extended with SDL_JoyBatteryEvent which has a field level of type SDL_JoystickPowerLevel which is defined in SDL_joystick.h. A number of functions defined in SDL_joystick.h need SDL_GUID which is defined in SDL_guid.h.

And down the rabbit hole we go...